requirements.txt
asyncio
captcha
cffi
cryptography
mysql-connector
Pillow
pycparser
PyYAML
uuid
websockets
jwt